# api/openapi.yaml
openapi: 3.1.0
info:
  title: PeerGit Platform API
  version: 0.1.0
  description: Health endpoints, identity and campus administration, media, and Phase 3 project/team/recruitment workflows.
servers:
  - url: https://peergit.localhost
paths:
...
  /api/v1/me/onboarding:
    get:
      operationId: campusOnboardingStatus
      summary: Read limited-account campus verification status and masked challenge metadata
      responses:
        '200': { description: Current user's onboarding status }
        '401': { description: Authentication required }
  /api/v1/me/campus-verification/challenges:
    post:
      operationId: createCampusEmailChallenge
      summary: Queue a ten-minute campus email link and OTP challenge
      responses:
        '201': { description: Challenge queued; response contains no credentials }
        '422': { description: Campus domain is not supported }
        '429': { description: Verification send limit or cooldown reached }
  /api/v1/me/campus-verification/confirm:
    post:
      operationId: confirmCampusEmail
      summary: Consume a link token or OTP with the current account's CSRF-protected session
      responses:
        '200': { description: Campus student access granted }
        '422': { description: Invalid, expired, superseded, or locked challenge }
...
  /api/v1/organizations:
    get:
      operationId: listOrganizations
      summary: List organizations for the authenticated campus
      security: [{ sessionCookie: [] }]
      responses:
        '200': { description: Tenant-scoped organizations }
        '403': { description: Campus scope or policy acceptance required }
  /api/v1/admin/invitations:
    post:
      operationId: createInvitation
      summary: Create a single-use campus or scoped-external invitation
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/InvitationCreate' }
      responses:
        '201': { description: Invitation created; URL returned once to the administrator }
        '403': { description: Campus administrator, consent, and recent MFA required }
  /api/v1/admin/invitations/{id}/revoke:
    post:
      operationId: revokeInvitation
      summary: Revoke an unused invitation in the current campus
      security: [{ sessionCookie: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/CsrfToken' }
      responses:
        '204': { description: Invitation revoked }
        '404': { description: Invitation not found, used, or belongs to another campus }
  /api/v1/admin/organizations:
    post:
      operationId: createOrganization
      summary: Create a tenant-owned campus organization
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OrganizationCreate' }
      responses:
        '201': { description: Organization created and audited }
        '409': { description: Slug already exists in the campus }
  /api/v1/admin/roles:
    post:
      operationId: grantCampusRole
      summary: Grant a non-administrator campus role to an active campus user
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: '#/components/parameters/CsrfToken' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoleGrant' }
      responses:
        '200': { description: Role granted and audited }
        '404': { description: Target account is outside the current campus }
  /api/v1/admin/users:
    get:
      operationId: listCampusUsers
      summary: Search a bounded page of current-campus users
      security: [{ sessionCookie: [] }]
      parameters:
        - { in: query, name: q, schema: { type: string, maxLength: 80 } }
        - { in: query, name: cursor, schema: { type: string, format: uuid } }
      responses:
        '200': { description: At most 50 tenant-scoped accounts and an opaque next ID }
  /api/v1/admin/users/{id}/status:
    post:
      operationId: changeCampusUserStatus
      summary: Suspend or reactivate a non-administrator campus account
      security: [{ sessionCookie: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
        - { $ref: '#/components/parameters/CsrfToken' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [suspended], properties: { suspended: { type: boolean } } }
      responses:
        '200': { description: Status changed, sessions revoked on suspension, and action audited }
        '409': { description: Self-suspension or administrator suspension requires break-glass review }
...
components:
  parameters:
    CsrfToken:
      in: header
      name: X-CSRF-Token
      required: true
      schema: { type: string }
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-peergit_session
  schemas:
...
    InvitationCreate:
      type: object
      required: [email, account_type]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        account_type: { type: string, enum: [campus, external] }
        role: { type: string, enum: [student, faculty, alumni_mentor, moderator] }
        external_access_kind: { type: string, enum: [mentor, recruiter, organization_guest] }
        organization_id: { type: string, format: uuid }
        expires_hours: { type: integer, minimum: 1, maximum: 168, default: 72 }
    OrganizationCreate:
      type: object
      required: [kind, slug, name]
      properties:
        kind: { type: string, enum: [department, club, innovation_cell, placement_cell, event_body] }
        slug: { type: string, pattern: '^[a-z0-9]+(-[a-z0-9]+)*$', minLength: 2, maxLength: 80 }
        name: { type: string, minLength: 2, maxLength: 160 }
        description: { type: string, maxLength: 2000 }
    RoleGrant:
      type: object
      required: [user_id, role]
      properties:
        user_id: { type: string, format: uuid }
        role: { type: string, enum: [student, faculty, alumni_mentor, moderator] }
...